Add Better Titles
Updating mods was really a hard thing to do when I had many tabs open
and couldn't find which was which. This simple change made it multiple
times better to organize my tabs.

Changing the title of the view can now be done by just adding the
section "title" to the page and adding the title tag inside of it.