URI encode build version to prevent special chars from being dropped on post
On update of the modpack recommended or latest builds, if the build version had a space or other special character the update would fail because the special characters were lost during the ajax post (even though the GUI presented a success message). By encoding the version number, the special characters are not lost during the ajax post.